rename testing/util -> testing/testutil
