+++ backend/schedulerd/schedulerd_test.go
@@ -8,73 +8,74 @@ import (
 
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/store/etcd"
-	"github.com/sensu/sensu-go/testing/util"
+	"github.com/sensu/sensu-go/testing/testutil"
 	"github.com/sensu/sensu-go/types"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestSchedulerd(t *testing.T) {
-	util.WithTempDir(func(tmpDir string) {
-		p := make([]int, 2)
-		err := util.RandomPorts(p)
-		if err != nil {
-			assert.FailNow(t, "failed to get ports for testing: ", err.Error())
-		}
+	tmpDir, remove := testutil.TempDir(t)
+	defer remove()
 
-		cfg := etcd.NewConfig()
-		cfg.DataDir = tmpDir
+	p := make([]int, 2)
+	err := testutil.RandomPorts(p)
+	if err != nil {
+		assert.FailNow(t, "failed to get ports for testing: ", err.Error())
+	}
 
-		peerURL := fmt.Sprintf("http://127.0.0.1:%d", p[1])
+	cfg := etcd.NewConfig()
+	cfg.DataDir = tmpDir
 
-		cfg.ListenClientURL = fmt.Sprintf("http://127.0.0.1:%d", p[0])
-		cfg.ListenPeerURL = peerURL
-		cfg.InitialCluster = fmt.Sprintf("default=http://127.0.0.1:%d", p[1])
-		cfg.InitialAdvertisePeerURL = peerURL
-		e, err := etcd.NewEtcd(cfg)
-		assert.NoError(t, err)
-		defer e.Shutdown()
+	peerURL := fmt.Sprintf("http://127.0.0.1:%d", p[1])
 
-		st, err := e.NewStore()
-		bus := &messaging.WizardBus{}
-		assert.NoError(t, bus.Start())
+	cfg.ListenClientURL = fmt.Sprintf("http://127.0.0.1:%d", p[0])
+	cfg.ListenPeerURL = peerURL
+	cfg.InitialCluster = fmt.Sprintf("default=http://127.0.0.1:%d", p[1])
+	cfg.InitialAdvertisePeerURL = peerURL
+	e, err := etcd.NewEtcd(cfg)
+	assert.NoError(t, err)
+	defer e.Shutdown()
 
-		// Mock a default organization
-		st.UpdateOrganization(
-			context.Background(),
-			&types.Organization{
-				Name: "default",
-			})
+	st, err := e.NewStore()
+	bus := &messaging.WizardBus{}
+	assert.NoError(t, bus.Start())
 
-		checker := &Schedulerd{
-			Store:      st,
-			MessageBus: bus,
-		}
-		checker.Start()
+	// Mock a default organization
+	st.UpdateOrganization(
+		context.Background(),
+		&types.Organization{
+			Name: "default",
+		})
 
-		ch := make(chan interface{}, 10)
-		assert.NoError(t, bus.Subscribe("subscription", "channel", ch))
+	checker := &Schedulerd{
+		Store:      st,
+		MessageBus: bus,
+	}
+	checker.Start()
 
-		check := types.FixtureCheckConfig("check_name")
-		ctx := context.WithValue(context.Background(), types.OrganizationKey, check.Organization)
+	ch := make(chan interface{}, 10)
+	assert.NoError(t, bus.Subscribe("subscription", "channel", ch))
 
-		assert.NoError(t, check.Validate())
-		assert.NoError(t, st.UpdateCheckConfig(ctx, check))
+	check := types.FixtureCheckConfig("check_name")
+	ctx := context.WithValue(context.Background(), types.OrganizationKey, check.Organization)
 
-		time.Sleep(1 * time.Second)
+	assert.NoError(t, check.Validate())
+	assert.NoError(t, st.UpdateCheckConfig(ctx, check))
 
-		err = st.DeleteCheckConfigByName(ctx, check.Name)
-		assert.NoError(t, err)
+	time.Sleep(1 * time.Second)
 
-		time.Sleep(1 * time.Second)
+	err = st.DeleteCheckConfigByName(ctx, check.Name)
+	assert.NoError(t, err)
 
-		assert.NoError(t, checker.Stop())
-		assert.NoError(t, bus.Stop())
-		close(ch)
+	time.Sleep(1 * time.Second)
 
-		for msg := range ch {
-			result, ok := msg.(*types.CheckConfig)
-			assert.True(t, ok)
-			assert.EqualValues(t, check, result)
-		}
-	})
+	assert.NoError(t, checker.Stop())
+	assert.NoError(t, bus.Stop())
+	close(ch)
+
+	for msg := range ch {
+		result, ok := msg.(*types.CheckConfig)
+		assert.True(t, ok)
+		assert.EqualValues(t, check, result)
+	}
 }
