+++ backend/store/etcd/etcd_test.go
@@ -9,67 +9,68 @@ import (
 	"testing"
 
 	"github.com/coreos/etcd/clientv3"
-	"github.com/sensu/sensu-go/testing/util"
+	"github.com/sensu/sensu-go/testing/testutil"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNewEtcd(t *testing.T) {
-	util.WithTempDir(func(tmpDir string) {
-		ports := make([]int, 2)
-		err := util.RandomPorts(ports)
-		if err != nil {
-			log.Panic(err)
-		}
-		clURL := fmt.Sprintf("http://127.0.0.1:%d", ports[0])
-		apURL := fmt.Sprintf("http://127.0.0.1:%d", ports[1])
-		initCluster := fmt.Sprintf("default=%s", apURL)
-		fmt.Println(initCluster)
+	tmpDir, remove := testutil.TempDir(t)
+	defer remove()
 
-		cfg := NewConfig()
-		cfg.DataDir = tmpDir
-		cfg.ListenClientURL = clURL
-		cfg.ListenPeerURL = apURL
-		cfg.InitialCluster = initCluster
-		cfg.InitialClusterState = ClusterStateNew
-		cfg.InitialAdvertisePeerURL = apURL
-		cfg.Name = "default"
+	ports := make([]int, 2)
+	err := testutil.RandomPorts(ports)
+	if err != nil {
+		log.Panic(err)
+	}
+	clURL := fmt.Sprintf("http://127.0.0.1:%d", ports[0])
+	apURL := fmt.Sprintf("http://127.0.0.1:%d", ports[1])
+	initCluster := fmt.Sprintf("default=%s", apURL)
+	fmt.Println(initCluster)
 
-		e, err := NewEtcd(cfg)
-		if e != nil {
-			defer e.Shutdown()
-		}
-		assert.NoError(t, err)
-		if err != nil {
-			pprof.Lookup("goroutine").WriteTo(os.Stdout, 1)
-			pprof.Lookup("threadcreate").WriteTo(os.Stdout, 1)
-			pprof.Lookup("heap").WriteTo(os.Stdout, 1)
-			assert.FailNow(t, "unable to start new etcd")
-		}
+	cfg := NewConfig()
+	cfg.DataDir = tmpDir
+	cfg.ListenClientURL = clURL
+	cfg.ListenPeerURL = apURL
+	cfg.InitialCluster = initCluster
+	cfg.InitialClusterState = ClusterStateNew
+	cfg.InitialAdvertisePeerURL = apURL
+	cfg.Name = "default"
 
-		client, err := e.NewClient()
-		assert.NoError(t, err)
-		kv := clientv3.NewKV(client)
-		assert.NotNil(t, kv)
+	e, err := NewEtcd(cfg)
+	if e != nil {
+		defer e.Shutdown()
+	}
+	assert.NoError(t, err)
+	if err != nil {
+		pprof.Lookup("goroutine").WriteTo(os.Stdout, 1)
+		pprof.Lookup("threadcreate").WriteTo(os.Stdout, 1)
+		pprof.Lookup("heap").WriteTo(os.Stdout, 1)
+		assert.FailNow(t, "unable to start new etcd")
+	}
 
-		putsResp, err := kv.Put(context.Background(), "key", "value")
-		assert.NoError(t, err)
-		assert.NotNil(t, putsResp)
+	client, err := e.NewClient()
+	assert.NoError(t, err)
+	kv := clientv3.NewKV(client)
+	assert.NotNil(t, kv)
 
-		if putsResp == nil {
-			assert.FailNow(t, "got nil put response from etcd")
-		}
+	putsResp, err := kv.Put(context.Background(), "key", "value")
+	assert.NoError(t, err)
+	assert.NotNil(t, putsResp)
 
-		getResp, err := kv.Get(context.Background(), "key")
-		assert.NoError(t, err)
-		assert.NotNil(t, getResp)
+	if putsResp == nil {
+		assert.FailNow(t, "got nil put response from etcd")
+	}
 
-		if getResp == nil {
-			assert.FailNow(t, "got nil get response from etcd")
-		}
-		assert.Equal(t, 1, len(getResp.Kvs))
-		assert.Equal(t, "key", string(getResp.Kvs[0].Key))
-		assert.Equal(t, "value", string(getResp.Kvs[0].Value))
+	getResp, err := kv.Get(context.Background(), "key")
+	assert.NoError(t, err)
+	assert.NotNil(t, getResp)
 
-		e.Shutdown()
-	})
+	if getResp == nil {
+		assert.FailNow(t, "got nil get response from etcd")
+	}
+	assert.Equal(t, 1, len(getResp.Kvs))
+	assert.Equal(t, "key", string(getResp.Kvs[0].Key))
+	assert.Equal(t, "value", string(getResp.Kvs[0].Value))
+
+	e.Shutdown()
 }
